Added an auto to just go into the observation zone
